requirements.txt
streamlit>=1.28.0
requests>=2.31.0
gitpython>=3.1.40
tree-sitter>=0.20.4
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-cpp>=0.22.0
tree-sitter-c>=0.21.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.0.0
typing-extensions>=4.8.0
